:ambulance: fix(Validations): Refactor child assent validator and fix gender check
Removed an unnecessary check for gender in `PreFlourishChildAssentFormValidator` class to simplify the code. The female gender check has been replaced by a more explicit evaluation in `validate_preg_testing` function, making it applicable when the participant's gender is female, and she is 12 years old or older. This makes the overall structure of the code cleaner and the gender specification for the pregnancy testing more accurate and explicit. Also, unnecessary blank lines have been removed for better code readability.