:sparkles: feature(Dashboard): Add age check to define Flourish eligibility
Three changes have been made to the application. Firstly, "age" function was added to the caregiver dashboard view. It calculates the age of the child and checks if it falls within the defined range (7-9.5 years) to determine Flourish eligibility. The "age" function was added as it offers a more flexible way to determine eligibility based on age rather than just BMI. Additionally, workflow steps have been updated to include a "Checkout code" step prior to the "repo-sync" step within the Sync Repo workflow. Lastly, the 'screening_identifier' check has been moved inside an 'if' block in the caregiver consent form to improve error handling.