:recycle: Refactor: Improve Flourish eligibility checks in caregiver dashboard view
Simplified and optimized the process for determining Flourish eligibility in the caregiver dashboard view. "get_utcnow" import statement was removed since it was not being used. The 'valid_by_age' and 'is_flourish_eligible' functions were refactored to return True as soon as the conditions for eligibility are met, rather than appending all results into a list and using the any() function