:recycle: refactor(Consent helper): Refactor subject registration and locator update logic
Removed the check for naive report time and timezone aware datetime logic in registration, simplifying the 'get_registered_subject_by_identifier_and_time' method in 'consent_helper.py'. This was done to reduce complexity as timezone handling is now handled uniformly elsewhere in the system.

In 'signals.py', revised the assignment of 'subject_identifier' and 'screening_identifier' in 'locator_obj' to only occur if they are not already assigned, enhancing the reliability of the locator object update process. This change ensures the locator keeps accurate tracking of subject's consent and screening status without overwriting existing data.