:sparkles: feat(Consent): Introduce consent version 4 checking in dashboard
This commit introduces the addition of a new parameter to check if the user has the latest consent version before rendering the "Consent Version" button in the dashboard. This improvement seeks to enhance user interaction by ensuring that outdated versions are not used. It also involves refactoring of some portions of the `dashboard_view_mixin.py` script towards improving readability and maintainability of the code.