(refactor): Tidy up code in pre_flourish_child_assent model
This commit includes cosmetic changes in the model file, pre_flourish_child_assent in the pre_flourish app. The unused newline is removed and the spaces for indentation have been made consistent. The verboseness of a question in the form fields is split into two lines for readability. Finally, fetching the latest consent instance from the database is now done more efficiently by using the latest command instead of get.

Signed-off-by: nmunatsibw <mcturnergarai@gmail.com>